refactor and fix up sponsor carousel
